+++ src/parser/argument-list.js
@@ -44,10 +44,10 @@ export default class ArgumentList extends Array {
 	/**
 	 * Generates an object containing the arguments for the help screen.
 	 *
-	 * @returns {Object}
+	 * @returns {Promise<Object>}
 	 * @access public
 	 */
-	generateHelp() {
+	async generateHelp() {
 		const entries = [];
 
 		for (const { desc, hidden, hint, multiple, name, required } of this) {